LayoutTests/imported/w3c/web-platform-tests/XMLHttpRequest/responsexml-basic-expected.txt


FAIL XMLHttpRequest: responseXML basic test assert_equals: getElementById("n2") expected null but got Element node <p xmlns="namespacesarejuststrings" id="n2">2</p>
PASS responseXML on empty response documents 
